Updated to version 1.1
I added an automatic value assignment. So just set the bone as you wish and this values will be used to setup the driver.
Also you now can create new shapes directly from the operator. And the operator is now located in the pose bones special menu and pose tool panel.